+++ patacrep/content/__init__.py
@@ -116,6 +116,10 @@ class ContentItem:
         """Return the string to end a block."""
         return ""
 
+    def file_entry(self):
+        """Return the dict representation (as in the yaml file)."""
+        raise NotImplementedError()
+
 class ContentError(SharedError):
     """Error in a content plugin."""
     def __init__(self, keyword=None, message=None):