+++ patacrep/content/setcounter.py
@@ -14,6 +14,9 @@ class CounterSetter(ContentItem):
         """Set the value of the counter."""
         return r'\setcounter{{{}}}{{{}}}'.format(self.name, self.value)
 
+    def file_entry(self):
+        return {'setcounter': {'name': self.name, 'value': self.value}}
+
 #pylint: disable=unused-argument
 @validate_parser_argument("""
 type: //any